// src/features/wallet/models/network/types/NetworkLoadState.ts
export enum NetworkLoadState {
  IDLE = 'IDLE',
  NETWORK_REQUESTED = 'NETWORK_REQUESTED',
  NETWORK_DETECTION_FAILED = 'NETWORK_DETECTION_FAILED',
  WRONG_NETWORK = 'WRONG_NETWORK',
  NETWORK_SWITCH_REQUESTED = 'NETWORK_SWITCH_REQUESTED',
  NETWORK_SWITCH_REJECTED = 'NETWORK_SWITCH_REJECTED',
  NETWORK_SWITCH_FAILED = 'NETWORK_SWITCH_FAILED',
  NETWORK_LOADED = 'NETWORK_LOADED',
}